Configuration for HICTOP 3DP08 Prusa i3 clone
This configuration appears to correctly drive the HICTOP 3DP08
Prusa i3 clone I have.  Since they have reportedly made changes
to their design from time to time, it might not work for your
printer, but it is likely to at least be a start.

I initially based this work on comparing the Configuration.h in the
Marlin sources they sent me.  It was not obvious which version or
versions of Marlin they based their firmware on.  I had to make
changes not present in that firmware source for it to work correctly.